+++ crates/cli/src/commands/install.rs
@@ -6,6 +6,7 @@ use crate::utils::tool_record::ToolRecord;
 use crate::workflows::{InstallOutcome, InstallWorkflowManager, InstallWorkflowParams};
 use clap::Args;
 use proto_core::flow::detect::Detector;
+use proto_core::flow::lock::Locker;
 use proto_core::{
     ConfigMode, Id, PinLocation, Tool, ToolContext, ToolSpec, reporter::NoticeOutput,
 };
@@ -171,6 +172,13 @@ pub async fn install_one(
         let config = session.load_config_with_mode(ConfigMode::UpwardsGlobal)?;
 
         enforce_requirements(&tool, &config.versions)?;
+
+        // Reconcile lockfiles by removing orphaned records, but keep
+        // records for the spec that is currently being installed
+        Locker::prune_orphaned_records(
+            &session.env,
+            &BTreeMap::from_iter([(context.clone(), spec.req.clone())]),
+        )?;
     }
 
     // Create our workflow and setup the progress reporter
@@ -297,6 +305,18 @@ async fn install_all(session: ProtoSession, args: InstallArgs) -> SessionResult 
         })?;
 
         return Ok(Some(1));
+    }
+
+    // Reconcile lockfiles by removing orphaned records, but keep
+    // records for the specs that are currently being installed
+    if !args.internal {
+        Locker::prune_orphaned_records(
+            &session.env,
+            &versions
+                .iter()
+                .map(|(context, spec)| (context.to_owned(), spec.req.to_owned()))
+                .collect(),
+        )?;
     }
 
     // Then install each tool in parallel!